Move drawing out of terminal
Move the drawing functionality out of the terminal and into each of the
objects being drawn.
