+++ tests/test_window.py
@@ -6,6 +6,7 @@ from seligimus.maths.integer_position_2d import IntegerPosition2D
 from seligimus.maths.integer_size_2d import IntegerSize2D
 
 from illud.buffer import Buffer
+from illud.canvas import Canvas
 from illud.exceptions.no_columns_exception import NoColumnsException
 from illud.exceptions.no_rows_exception import NoRowsException
 from illud.window import Window
@@ -227,3 +228,43 @@ def test_adjust_view_to_include(window: Window, index: int, expected_window: Win
     window.adjust_view_to_include(index)
 
     assert window == expected_window
+
+
+# yapf: disable # pylint: disable=line-too-long
+@pytest.mark.parametrize('window, canvas, expected_canvas_after', [
+    (Window(), Canvas(), Canvas()),
+    (Window(size=IntegerSize2D(1, 1)), Canvas(IntegerSize2D(1, 1), [['x']]), Canvas(IntegerSize2D(1, 1), [[' ']])),
+    (Window(size=IntegerSize2D(1, 1), buffer_=Buffer('a')), Canvas(IntegerSize2D(1, 1), [['x']]), Canvas(IntegerSize2D(1, 1), [['a']])),
+    (Window(size=IntegerSize2D(2, 1), buffer_=Buffer('foo')), Canvas(IntegerSize2D(2, 1), [['x', 'x']]), Canvas(IntegerSize2D(2, 1), [['f', 'o']])),
+    (Window(size=IntegerSize2D(3, 1), buffer_=Buffer('foo')), Canvas(IntegerSize2D(3, 1), [['x', 'x', 'x']]), Canvas(IntegerSize2D(3, 1), [['f', 'o', 'o']])),
+    (Window(size=IntegerSize2D(2, 1), buffer_=Buffer('foo')), Canvas(IntegerSize2D(3, 1), [['x', 'x', 'x']]), Canvas(IntegerSize2D(3, 1), [['f', 'o', 'x']])),
+    (Window(size=IntegerSize2D(3, 1), buffer_=Buffer('foo')), Canvas(IntegerSize2D(4, 1), [['x', 'x', 'x', 'x']]), Canvas(IntegerSize2D(4, 1), [['f', 'o', 'o', 'x']])),
+    (Window(size=IntegerSize2D(3, 1), buffer_=Buffer('foo')), Canvas(IntegerSize2D(5, 1), [['x', 'x', 'x', 'x', 'x']]), Canvas(IntegerSize2D(5, 1), [['f', 'o', 'o', 'x', 'x']])),
+    (Window(size=IntegerSize2D(5, 1), buffer_=Buffer('foo')), Canvas(IntegerSize2D(5, 1), [['x', 'x', 'x', 'x', 'x']]), Canvas(IntegerSize2D(5, 1), [['f', 'o', 'o', ' ', ' ']])),
+    (Window(size=IntegerSize2D(5, 1), buffer_=Buffer('foo\n')), Canvas(IntegerSize2D(5, 1), [['x', 'x', 'x', 'x', 'x']]), Canvas(IntegerSize2D(5, 1), [['f', 'o', 'o', ' ', ' ']])),
+    (Window(size=IntegerSize2D(2, 2), buffer_=Buffer('foo')), Canvas(IntegerSize2D(3, 2), [['x', 'x', 'x'], ['x', 'x', 'x']]), Canvas(IntegerSize2D(3, 2), [['f', 'o', 'x'], [' ', ' ', 'x']])),
+    (Window(size=IntegerSize2D(1, 2), buffer_=Buffer('f')), Canvas(IntegerSize2D(2, 2), [['x', 'x'], ['x', 'x']]), Canvas(IntegerSize2D(2, 2), [['f', 'x'], [' ', 'x']])),
+    (Window(size=IntegerSize2D(1, 2), buffer_=Buffer('f\nb')), Canvas(IntegerSize2D(2, 2), [['x', 'x'], ['x', 'x']]), Canvas(IntegerSize2D(2, 2), [['f', 'x'], ['b', 'x']])),
+    (Window(size=IntegerSize2D(1, 2), buffer_=Buffer('foo\nb')), Canvas(IntegerSize2D(2, 2), [['x', 'x'], ['x', 'x']]), Canvas(IntegerSize2D(2, 2), [['f', 'x'], ['b', 'x']])),
+    (Window(size=IntegerSize2D(1, 2), buffer_=Buffer('foo\nbar')), Canvas(IntegerSize2D(2, 2), [['x', 'x'], ['x', 'x']]), Canvas(IntegerSize2D(2, 2), [['f', 'x'], ['b', 'x']])),
+    (Window(size=IntegerSize2D(2, 2), buffer_=Buffer('f\nb')), Canvas(IntegerSize2D(2, 2), [['x', 'x'], ['x', 'x']]), Canvas(IntegerSize2D(2, 2), [['f', ' '], ['b', ' ']])),
+    (Window(size=IntegerSize2D(5, 2), buffer_=Buffer('foo\nbar')), Canvas(IntegerSize2D(5, 2), [['x', 'x', 'x', 'x', 'x'], ['x', 'x', 'x', 'x', 'x']]), Canvas(IntegerSize2D(5, 2), [['f', 'o', 'o', ' ', ' '], ['b', 'a', 'r', ' ', ' ']])),
+    (Window(size=IntegerSize2D(5, 3), buffer_=Buffer('foo\nbar\nbaz')), Canvas(IntegerSize2D(5, 3), [['x', 'x', 'x', 'x', 'x'], ['x', 'x', 'x', 'x', 'x'], ['x', 'x', 'x', 'x', 'x']]), Canvas(IntegerSize2D(5, 3), [['f', 'o', 'o', ' ', ' '], ['b', 'a', 'r', ' ', ' '], ['b', 'a', 'z', ' ', ' ']])),
+    (Window(size=IntegerSize2D(3, 1), buffer_=Buffer('foo'), offset=IntegerPosition2D(1, 0)), Canvas(IntegerSize2D(3, 1), [['x', 'x', 'x']]), Canvas(IntegerSize2D(3, 1), [['o', 'o', ' ']])),
+    (Window(size=IntegerSize2D(3, 2), buffer_=Buffer('foo\nbar'), offset=IntegerPosition2D(1, 0)), Canvas(IntegerSize2D(3, 2), [['x', 'x', 'x'], ['x', 'x', 'x']]), Canvas(IntegerSize2D(3, 2), [['o', 'o', ' '], ['a', 'r', ' ']])),
+    (Window(size=IntegerSize2D(3, 3), buffer_=Buffer('foo\n\nbar'), offset=IntegerPosition2D(1, 0)), Canvas(IntegerSize2D(3, 3), [['x', 'x', 'x'], ['x', 'x', 'x'], ['x', 'x', 'x']]), Canvas(IntegerSize2D(3, 3), [['o', 'o', ' '], [' ', ' ', ' '], ['a', 'r', ' ']])),
+    (Window(size=IntegerSize2D(3, 1), buffer_=Buffer('foo'), offset=IntegerPosition2D(-1, 0)), Canvas(IntegerSize2D(3, 1), [['x', 'x', 'x']]), Canvas(IntegerSize2D(3, 1), [[' ', 'f', 'o']])),
+    (Window(size=IntegerSize2D(3, 2), buffer_=Buffer('foo\nbar'), offset=IntegerPosition2D(-1, 0)), Canvas(IntegerSize2D(3, 2), [['x', 'x', 'x'], ['x', 'x', 'x']]), Canvas(IntegerSize2D(3, 2), [[' ', 'f', 'o'], [' ', 'b', 'a']])),
+    (Window(size=IntegerSize2D(3, 3), buffer_=Buffer('foo\n\nbar'), offset=IntegerPosition2D(-1, 0)), Canvas(IntegerSize2D(3, 3), [['x', 'x', 'x'], ['x', 'x', 'x'], ['x', 'x', 'x']]), Canvas(IntegerSize2D(3, 3), [[' ', 'f', 'o'], [' ', ' ', ' '], [' ', 'b', 'a']])),
+    (Window(size=IntegerSize2D(3, 1), buffer_=Buffer('foo'), offset=IntegerPosition2D(-4, 0)), Canvas(IntegerSize2D(3, 1), [['x', 'x', 'x']]), Canvas(IntegerSize2D(3, 1), [[' ', ' ', ' ']])),
+    (Window(size=IntegerSize2D(3, 2), buffer_=Buffer('foo'), offset=IntegerPosition2D(0, 1)), Canvas(IntegerSize2D(3, 2), [['x', 'x', 'x'], ['x', 'x', 'x']]), Canvas(IntegerSize2D(3, 2), [[' ', ' ', ' '], [' ', ' ', ' ']])),
+    (Window(size=IntegerSize2D(3, 2), buffer_=Buffer('foo'), offset=IntegerPosition2D(0, -1)), Canvas(IntegerSize2D(3, 2), [['x', 'x', 'x'], ['x', 'x', 'x']]), Canvas(IntegerSize2D(3, 2), [[' ', ' ', ' '], ['f', 'o', 'o']])),
+    (Window(size=IntegerSize2D(3, 2), buffer_=Buffer('foo\nbar'), offset=IntegerPosition2D(0, -1)), Canvas(IntegerSize2D(3, 2), [['x', 'x', 'x'], ['x', 'x', 'x']]), Canvas(IntegerSize2D(3, 2), [[' ', ' ', ' '], ['f', 'o', 'o']])),
+    (Window(size=IntegerSize2D(3, 2), buffer_=Buffer('foo'), offset=IntegerPosition2D(0, -2)), Canvas(IntegerSize2D(3, 2), [['x', 'x', 'x'], ['x', 'x', 'x']]), Canvas(IntegerSize2D(3, 2), [[' ', ' ', ' '], [' ', ' ', ' ']])),
+])
+# yapf: enable # pylint: enable=line-too-long
+def test_draw(window: Window, canvas: Canvas, expected_canvas_after: Canvas) -> None:
+    """Test illud.window.Window.draw."""
+    window.draw(canvas)
+
+    assert canvas == expected_canvas_after
